Tue, Nov 8, 2022, 7:29 AM +00:00

+++ 2_Array/problem/readability/readability.c
@@ -99,8 +99,8 @@ int compute_coleman_liau(int letters, int words, int sentences)
     // wordsで割る＝1単語あたりの数を出して100倍するだけだった
     // wikipediaでちょっとカンニングした
     // https://en.wikipedia.org/wiki/Coleman%E2%80%93Liau_index
-    float l = (letters / words) * 100.0;
-    float s = (sentences / words) * 100.0;
+    float l = (letters / (float)words) * 100;
+    float s = (sentences / (float)words) * 100;
     int grade = round(0.0588 * l - 0.296 * s - 15.8);
     return grade;
 }
